requirements.txt
streamlit==1.30.0
streamlit_lottie==0.0.5
Pillow==10.2.0
requests==2.31.0